improved orchidea.dbquery external and help

+++ max/src/orchidea_Max/main/orchidea.features.cpp
@@ -157,7 +157,8 @@ void* orchidea_dispatcher (void* d) {
             for (long i = 0; i < x->filenames_size; i++) {
                 std::vector<float> features;
                 if (x->filenames[i]) {
-                    compute_features<float> (x->filenames[i]->s_name, features, x->win_size, x->hop_size, 4, "moments");
+                    compute_features<float> (x->filenames[i]->s_name, features,
+                                             (int) x->win_size, (int) x->hop_size, 4, "moments");
                     
                     //        compute_features<float> (x->file_name->s_name, features,
                     //                                 x->win_size, x->hop_size, 4, "transients");
